+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.0.40
 
-- ...
+- Fixed specific problematic 'alt+n' keybinding for macOS when Colemak keyboard layout is used.
 
 ## 1.0.39